Merges from release-0.5.3-branch

+++ deletedomain.php
@@ -7,6 +7,10 @@
 if ($admintype==0){
 
 	$handle=DB::connect($DSN, true);
+	if (DB::isError($handle)) {
+		die (_("Database error"));
+	}
+	
 	$query1="SELECT * FROM accountuser WHERE domain_name='$domain' order by username";
 	$result1=$handle->query($query1);
 	$cnt1=$result1->numRows();
@@ -17,10 +21,13 @@ if ($admintype==0){
 		?>
 		<h3><?php print _("Delete a Domain from the System") ?></h3>
 
-		<h3><?php print _("Do you really want to delete the Domain")?> <font color=red><?php print $domain ?></font> <?php print _("with all its defined accounts, admins, and emailadresses")?>?</h3>
+		<h3><?php print _("Do you really want to delete the Domain")?> 
+		<font color=red><?php print $domain ?></font> 
+		<?php print _("with all its defined accounts, admins, and emailadresses")?>?</h3>
 		<?php print _("This can take a while depending on how many account have to be deleted")?><p>
 
-		<font color="red"><?php print _("Your action will delete")." " ?> <?php print $cnt1 ."&nbsp;"._("accounts")?> </font><p>
+		<font color="red"><?php print _("Your action will delete")." " ?> 
+		<?php print $cnt1 ."&nbsp;"._("accounts")?> </font><p>
 
 		<form action="index.php">
 		<input type="hidden" name="action" value="deletedomain">
@@ -46,6 +53,7 @@ if ($admintype==0){
 	        $cyr_conn -> imap_login();
 
 		# First Delete all stuff related to the domain from the database
+
 	
 		$query2="DELETE FROM virtual WHERE domain_name='$domain'";
 		$hnd2=$handle->query($query2);
@@ -73,19 +81,48 @@ if ($admintype==0){
 			}
 
 		}
+
+		# Finally the domain must be removed from the domainadmin table
 	
 		$query6="SELECT * FROM domainadmin WHERE domain_name='$domain'";
 		$result6=$handle->query($query6);
 		$cnt6=$result6->numRows();
-                for ($i=0;$i<$cnt6;$i++){
+                for ($i=0;$i<=$cnt6;$i++){
+
+			# After getting the resulttable we search for the adminuser 
+			# in each row
+
 			$row=$result6->fetchRow($i);
-			$username=$row['adminuser'];
-                        $query7="DELETE FROM adminuser where username='$username'";
-                        $result7=$handle->query($query7);
+			$username=$row['1'];
+			$query7="SELECT * FROM domainadmin where adminuser='$username'";
+			$result7=$handle->query($query7);
+			$cnt7=$result7->numRows();
+
+			# If the adminuser is only the admin for the domain to be deleted,
+			# then this adminuser also needs to be deleted
+
+			if ($cnt7==1){
+                        	$query7="DELETE FROM adminuser where username='$username'";
+				$result7=$handle->query($query7);
+			}
+
+			# Finally delete every entry with the  domain to be deleted
+	
+                        $query8="DELETE FROM domainadmin where domain_name='$domain'";
+                        $result8=$handle->query($query8);
                 } 
-		$query8="DELETE FROM domainadmin WHERE domain_name='$domain'";
-                $hnd8=$handle->query($query8); 
 	print "<h3>". _("Domain")." ".$domain." ". _("successfully deleted")."</h3>";
+
+	unset ($domain);
+
+?>
+<script type="text/javascript">
+<!--
+ window.location.href = "index.php";
+//-->
+
+</script>
+<?php
 
 	include ("browse.php");
